Add more system tests: class testing in particular
