+++ .gitignore
@@ -2,3 +2,4 @@ pkg/
 Gemfile.lock
 # TODO: Ignore this for now until we decide what to do with it
 spec/fixtures/manifests/
+.ruby-version
