+++ .nodeset.yml
@@ -3,29 +3,29 @@ default_set: 'centos-64-x64'
 sets:
   'centos-59-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'centos-59-x64'
   'centos-64-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'centos-64-x64'
   'fedora-18-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'fedora-18-x64'
   'debian-607-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'debian-607-x64'
   'debian-70rc1-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'debian-70rc1-x64'
   'ubuntu-server-10044-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'ubuntu-server-10044-x64'
   'ubuntu-server-12042-x64':
     nodes:
-      "main":
+      "main.foo.vm":
         prefab: 'ubuntu-server-12042-x64'
