+++ spec/spec_helper_system.rb
@@ -3,9 +3,34 @@
 require 'rspec-system/spec_helper'
 require 'rspec-system-puppet/helpers'
 
+# Just some helpers specific to this module
+module LocalHelpers
+   # This helper flushes all tables on the default machine.
+   #
+   # It checks that the flush command returns with no errors.
+   #
+   # @return [void]
+   # @todo Need to optionally do the newer tables
+   # @example
+   #   it 'should flush tables' do
+   #     iptables_flush_all_tables
+   #   end
+   def iptables_flush_all_tables
+     ['filter', 'nat', 'mangle', 'raw'].each do |t|
+       system_run("/sbin/iptables -t #{t} -F") do |r|
+         r[:exit_code].should == 0
+         r[:stderr].should == ''
+       end
+     end
+   end
+end
+
 RSpec.configure do |c|
   # Project root for the firewall code
   proj_root = File.expand_path(File.join(File.dirname(__FILE__), '..'))
+
+  # Import in our local helpers
+  c.include ::LocalHelpers
 
   # This is where we 'setup' the nodes before running our tests
   c.system_setup_block = proc do
